Moving patplan and payplan to crud.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@4725 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms/FormImportXML.cs
@@ -434,7 +434,7 @@ namespace OpenDental{
 			//Then attach plan
 			List <PatPlan> PatPlanList=PatPlans.Refresh(pat.PatNum);
 			PatPlan patplan=new PatPlan();
-			patplan.Ordinal=PatPlanList.Count+1;//so the ordinal of the first entry will be 1, NOT 0.
+			patplan.Ordinal=(byte)(PatPlanList.Count+1);//so the ordinal of the first entry will be 1, NOT 0.
 			patplan.PatNum=pat.PatNum;
 			patplan.PlanNum=plan.PlanNum;
 			switch(insRelat){
